Trigger joining in the lifecycle saga
I make a helper out of the join.saga and trigger it as a part of the lifecycle. Glad it got nicified recently. XD
